OE-259 add year_multiplier

+++ database/migrations/2021_05_12_194316_create_stock_points_calculation_bases_table.php
@@ -11,7 +11,7 @@ class CreateStockPointsCalculationBasesTable extends Migration
         Schema::create('stock_points_calculation_bases', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('stock_base_point');
+            $table->decimal('stock_base_point');
             $table->timestamps();
         });
     }
